OSAL: confine every syscall to osal.c, and add Windows and macOS support
The library had socket code in two places: osal.c and dbserver.c, plus a poll()
loop in context.c. All of it now lives behind the OSAL, so bringing up a new
platform means editing one file:

  djl_sock_poll        readiness for the four DJ Link sockets (poll / select)
  djl_tcp_connect      bounded-wait connect, then blocking transfer
  djl_tcp_send_all     full-write and full-read helpers for dbserver
  djl_tcp_recv_exact

dbserver.c and context.c no longer include a single system header, which is
what actually made the port possible.

Windows: Winsock2 for sockets, IP Helper (GetAdaptersAddresses) for interface
lookup, matching on adapter name, friendly name or description since Windows
has no stable short names, QueryPerformanceCounter for the monotonic clock, and
select() where POSIX uses poll(). Winsock is started once, lazily.

macOS and BSD: the hardware address comes from AF_LINK / sockaddr_dl instead of
Linux's AF_PACKET, SO_BINDTODEVICE is absent so the interface filter is skipped
there, and SO_NOSIGPIPE replaces MSG_NOSIGNAL.

Socket handles are now carried as intptr_t rather than int. A Windows SOCKET is
a UINT_PTR and does not fit in an int on 64-bit; intptr_t round-trips it
exactly, INVALID_SOCKET lands on -1, and a POSIX descriptor is unchanged, so
"fd < 0 means invalid" still holds everywhere. A single DJL_SOCK macro converts
back to the native type at each syscall.

Verified, not just compiled: the whole library cross-compiles with mingw-w64
and the Windows binaries run under Wine, where all 851 checks pass and both
socket paths work against the live rig. The Windows NFS client mounts
CDJ-3000X player 1 and returns byte-identical results to the Linux build (482
beats, 2 cues, 13 phrases), Windows interface enumeration finds the adapter and
receives DJ Link broadcasts to discover the player, and the Windows dbserver
client fetches track 33 over TCP. macOS is written to the documented BSD API
but is not compiled here, since there is no SDK on this machine.

Also turns on the rest of the warning set ARCHITECTURE.md section 13 mandates
(-Wconversion -Wcast-qual -Wundef) now that the tree is clean under it, and
fixes what that exposed: sign-conversion in the dbserver beat-grid and le32
readers, which only showed up at -O0, an enum-to-uint8_t narrowing in the
context setup, and two cosmetic conversions in the CLI tools.

Builds and passes in four configurations: Linux release (851), Linux
ASan/UBSan (851), core with DJL_WITH_NFS=OFF (765, since the XDR and RPC tests
are compiled out with the client), and Windows via mingw-w64 (851).

+++ tools/djl_monitor.c
@@ -179,7 +179,7 @@ static void handle(const djl_event *ev, djl_context *ctx)
                m->track_count, m->playlist_count,
                djl_track_type_name(m->track_type),
                m->has_my_settings ? "yes" : "no",
-               m->free_bytes / 1e9, m->total_bytes / 1e9);
+               (double)m->free_bytes / 1e9, (double)m->total_bytes / 1e9);
         break;
     }
     case DJL_EV_CDJ_STATUS: {
